Python/usb_prog/test_usb_prog.py
import unittest
from unittest.mock import MagicMock, patch
import usb.core
import usb.util

# Assuming the functions are imported from your main script
# from your_module import find_device, detach_kernel_driver, set_device_configuration, send_command, receive_data, reattach_kernel_driver

class TestUSBDeviceFunctions(unittest.TestCase):

    @patch('usb.core.find')
    def test_find_device_success(self, mock_find):
        # Mock the device being found
        mock_device = MagicMock()
        mock_find.return_value = mock_device

        device = find_device(0x1234, 0x5678)
        self.assertIsNotNone(device)
        self.assertEqual(device, mock_device)
        mock_find.assert_called_with(idVendor=0x1234, idProduct=0x5678)

    @patch('usb.core.find')
    def test_find_device_failure(self, mock_find):
        # Mock the device not being found
        mock_find.return_value = None

        with self.assertRaises(ValueError):
            find_device(0x1234, 0x5678)

    def test_detach_kernel_driver(self):
        mock_device = MagicMock()
        mock_device.is_kernel_driver_active.return_value = True

        detach_kernel_driver(mock_device)
        mock_device.is_kernel_driver_active.assert_called_with(0)
        mock_device.detach_kernel_driver.assert_called_with(0)

    def test_detach_kernel_driver_not_active(self):
        mock_device = MagicMock()
        mock_device.is_kernel_driver_active.return_value = False

        detach_kernel_driver(mock_device)
        mock_device.is_kernel_driver_active.assert_called_with(0)
        mock_device.detach_kernel_driver.assert_not_called()

    def test_set_device_configuration(self):
        mock_device = MagicMock()

        set_device_configuration(mock_device)
        mock_device.set_configuration.assert_called_once()

    def test_send_command(self):
        mock_device = MagicMock()

        endpoint_out = 0x01
        command = b"\x01\x02\x03\x04"
        send_command(mock_device, endpoint_out, command)
        mock_device.write.assert_called_with(endpoint_out, command)

    def test_receive_data(self):
        mock_device = MagicMock()
        expected_data = b"\x05\x06\x07\x08"
        mock_device.read.return_value = expected_data

        endpoint_in = 0x81
        size = 64
        data = receive_data(mock_device, endpoint_in, size)
        mock_device.read.assert_called_with(endpoint_in, size)
        self.assertEqual(data, expected_data)

    def test_reattach_kernel_driver(self):
        mock_device = MagicMock()
        mock_device.is_kernel_driver_active.return_value = False

        reattach_kernel_driver(mock_device)
        mock_device.is_kernel_driver_active.assert_called_with(0)
        mock_device.attach_kernel_driver.assert_called_with(0)

    def test_reattach_kernel_driver_already_active(self):
        mock_device = MagicMock()
        mock_device.is_kernel_driver_active.return_value = True

        reattach_kernel_driver(mock_device)
        mock_device.is_kernel_driver_active.assert_called_with(0)
        mock_device.attach_kernel_driver.assert_not_called()

if __name__ == "__main__":
    unittest.main()
